Adapted project for deployment at Vercel, with serveless functions
