+++ .gitignore
@@ -1,0 +1,4 @@
+.vercel
+*.env
+node_modules/
+dist/*
