--- backend/.env.example
@@ -1,4 +1,0 @@
-# Environment variables for URL Shortener
-MONGODB_URI=mongodb://localhost:27017/url_shortener
-SESSION_SECRET=your_session_secret_here
-PORT=3000
