--- backend/models/user.js
@@ -1,10 +1,0 @@
-import mongoose from 'mongoose';
-
-const userSchema = new mongoose.Schema({
-    username: { type: String, required: true, unique: true },
-    password: { type: String, required: true },
-});
-
-const User = mongoose.model('User', userSchema);
-
-export default User;
