--- backend/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "url-shortener-backend",
-  "version": "1.0.0",
-  "type": "module",
-  "dependencies": {
-    "bcryptjs": "^3.0.2",
-    "body-parser": "^2.2.0",
-    "connect-flash": "^0.1.1",
-    "cors": "^2.8.5",
-    "dotenv": "^16.5.0",
-    "express": "^5.1.0",
-    "express-session": "^1.18.1",
-    "mongoose": "^8.15.1",
-    "passport": "^0.7.0",
-    "passport-local": "^1.0.0",
-    "shortid": "^2.2.17",
-    "validator": "^13.15.15"
-  },
-  "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js",
-    "test": "echo \"No test specified\" && exit 0"
-  },
-  "devDependencies": {
-    "nodemon": "^3.1.10"
-  }
-}
