--- backend/routes/auth.js
@@ -1,81 +1,0 @@
-import express from "express";
-import passport from "passport";
-import bcrypt from "bcryptjs";
-import User from "../models/user.js";
-
-const router = express.Router();
-
-// Register
-router.post("/register", async (req, res) => {
-  try {
-    const { username, password } = req.body;
-
-    if (!username || !password) {
-      return res
-        .status(400)
-        .json({ error: "Username and password are required" });
-    }
-
-    const existingUser = await User.findOne({ username });
-    if (existingUser) {
-      return res.status(400).json({ error: "Username already exists" });
-    }
-
-    const hashedPassword = await bcrypt.hash(password, 10);
-    const user = new User({
-      username,
-      password: hashedPassword,
-    });
-
-    await user.save();
-
-    req.login(user, (err) => {
-      if (err) {
-        return res
-          .status(500)
-          .json({ error: "Error logging in after registration" });
-      }
-      return res.json({ user: { username: user.username } });
-    });
-  } catch (error) {
-    console.error("Registration error:", error);
-    res.status(500).json({ error: "Internal server error" });
-  }
-});
-
-// Login
-router.post("/login", (req, res, next) => {
-  passport.authenticate("local", (err, user, info) => {
-    if (err) {
-      return res.status(500).json({ error: "Internal server error" });
-    }
-    if (!user) {
-      return res
-        .status(401)
-        .json({ error: info.message || "Invalid credentials" });
-    }
-    req.login(user, (err) => {
-      if (err) {
-        return res.status(500).json({ error: "Error logging in" });
-      }
-      return res.json({ user: { username: user.username } });
-    });
-  })(req, res, next);
-});
-
-// Logout
-router.post("/logout", (req, res) => {
-  req.logout(() => {
-    res.json({ message: "Logged out successfully" });
-  });
-});
-
-// Get Authenticated User
-router.get("/user", (req, res) => {
-  if (!req.isAuthenticated()) {
-    return res.status(401).json({ error: "Not authenticated" });
-  }
-  res.json({ user: { username: req.user.username } });
-});
-
-export default router;
