--- frontend/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "url-shortener-frontend",
-  "version": "1.0.0",
-  "type": "module",
-  "dependencies": {
-    "@tailwindcss/vite": "^4.1.10",
-    "@vueuse/core": "^13.3.0",
-    "axios": "^1.10.0",
-    "flowbite": "^3.1.2",
-    "pinia": "^2.3.1",
-    "tailwindcss": "^4.1.10",
-    "vue": "^3.2.0",
-    "vue-chartjs": "^5.3.2",
-    "vue-router": "^4.5.1"
-  },
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "serve": "vite preview"
-  },
-  "devDependencies": {
-    "@vitejs/plugin-vue": "^5.2.4",
-    "vite": "^6.3.5"
-  }
-}
