+++ package.json
@@ -1,0 +1,42 @@
+{
+  "name": "url-shortener-monorepo",
+  "version": "1.0.0",
+  "type": "module",
+  "dependencies": {
+    "bcryptjs": "^3.0.2",
+    "body-parser": "^2.2.0",
+    "connect-flash": "^0.1.1",
+    "cors": "^2.8.5",
+    "dotenv": "^16.5.0",
+    "express": "^5.1.0",
+    "express-session": "^1.18.1",
+    "mongoose": "^8.15.1",
+    "passport": "^0.7.0",
+    "passport-local": "^1.0.0",
+    "shortid": "^2.2.17",
+    "validator": "^13.15.15",
+    "@tailwindcss/vite": "^4.1.10",
+    "@vueuse/core": "^13.3.0",
+    "axios": "^1.10.0",
+    "flowbite": "^3.1.2",
+    "pinia": "^2.3.1",
+    "tailwindcss": "^4.1.10",
+    "vue": "^3.2.0",
+    "vue-chartjs": "^5.3.2",
+    "vue-router": "^4.5.1",
+    "jsonwebtoken": "^9.0.0"
+  },
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
+    "backend:start": "node server.js",
+    "backend:dev": "nodemon server.js",
+    "test": "echo \"No test specified\" && exit 0"
+  },
+  "devDependencies": {
+    "nodemon": "^3.1.10",
+    "@vitejs/plugin-vue": "^5.2.4",
+    "vite": "^6.3.5"
+  }
+}
